server/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0

# Google Cloud services
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0
googlemaps==4.10.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Firebase Admin SDK
firebase-admin==6.2.0

# Email services
sendgrid==6.10.0

# HTTP and async
httpx==0.25.2
aiofiles==23.2.1

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
